+++ ReleaseNotes.md
@@ -1,3 +1,10 @@
+#1.8.0:
+* Updated to v2 of surrogate
+* Updated to newer ID package
+* Modified artifact version in surrogate to include timestamp
+* Added support to generate MEA files for developer use only
+* Added a README with some compiling and general information.
+
 #1.7.0: 
 Development still.
 * Updated code based on code reviews.